Fixed broken tests
