+++ app/src/androidTest/java/com/github/se/polyfit/ui/components/card/IngredientInfoCardTest.kt
@@ -31,7 +31,7 @@ class IngredientInfoCardTest {
 
       composeTestRule.onNodeWithText(Ingredient.default().name).assertIsDisplayed()
 
-      composeTestRule.onNodeWithText(Ingredient.default().amountFormated()).assertIsDisplayed()
+      composeTestRule.onNodeWithText(Ingredient.default().amountFormatted()).assertIsDisplayed()
     }
   }
 }
