+++ app/src/main/java/com/github/se/polyfit/MainActivity.kt
@@ -50,7 +50,7 @@ class MainActivity : ComponentActivity() {
       PolyfitTheme {
         val navController = rememberNavController()
         val navigation = Navigation(navController)
-        NavHost(navController = navController, startDestination = Route.RecipeRecFlow) {
+        NavHost(navController = navController, startDestination = Route.Register) {
           composable(Route.Map) {
             GenericScreen(
                 navController = navController,
